refactor: move depositFactoryABI, factoryABI to interfaces/
