+++ packages/portfolio-contract/src/utils/evm-orch-factory.ts
@@ -2,14 +2,14 @@
  * @file Utilities for using Factory.sol to emulate orchestration
  * `chain.makeAccount()` behavior on EVM chains (CREATE2 prediction, etc.).
  */
-import {
-  PermitTransferFromComponents,
-  PermitTransferFromInternalTypeName,
-} from '@agoric/orchestration/src/utils/permit2.ts';
 import { assert } from '@endo/errors';
 import { keccak_256 as keccak256 } from '@noble/hashes/sha3';
-import type { Abi, AbiParameterToPrimitiveType, Hex } from 'viem';
+import type { AbiParameterToPrimitiveType, Hex } from 'viem';
 import { computeCreate2Address, hashInitCode } from './create2.ts';
+import {
+  depositFactoryABI,
+  walletMulticallABI,
+} from '../interfaces/orch-factory.ts';
 
 // XXX: refactor EVMInterface to use { type: 'address', name: '...' }
 type FactoryI = {
@@ -53,81 +53,6 @@ export const predictWalletAddress = ({
   return out;
 };
 
-/**
- * @see {@link https://github.com/agoric-labs/agoric-to-axelar-local/blob/main/packages/axelar-local-dev-cosmos/src/contracts/Factory.sol Factory.sol (_execute method)}
- * 1520650 2026-01-24
- */
-export const factoryABI = [
-  {
-    type: 'constructor',
-    inputs: [
-      { name: 'gateway_', type: 'address' },
-      { name: 'gasReceiver_', type: 'address' },
-    ],
-    stateMutability: 'payable',
-  },
-  {
-    type: 'function',
-    name: 'createSmartWallet',
-    inputs: [{ name: 'expectedWalletAddress', type: 'address' }],
-    outputs: [],
-    stateMutability: 'nonpayable',
-  },
-  {
-    type: 'event',
-    name: 'SmartWalletCreated',
-    anonymous: false,
-    inputs: [
-      { name: 'wallet', type: 'address', indexed: true },
-      { name: 'owner', type: 'string', indexed: false },
-      { name: 'sourceChain', type: 'string', indexed: false },
-    ],
-  },
-] as const satisfies Abi;
-
-/**
- * @see {@link https://github.com/agoric-labs/agoric-to-axelar-local/blob/main/packages/axelar-local-dev-cosmos/src/contracts/DepositFactory.sol DepositFactory.sol}
- * 1520650 2026-01-24
- */
-export const depositFactoryABI = [
-  {
-    type: 'constructor',
-    inputs: [
-      { name: 'gateway_', type: 'address' },
-      { name: 'gasReceiver_', type: 'address' },
-      { name: 'permit2_', type: 'address' },
-      { name: 'owner_', type: 'string' },
-    ],
-    stateMutability: 'payable',
-  },
-  {
-    type: 'function',
-    name: 'createAndDeposit',
-    inputs: [
-      {
-        type: 'tuple',
-        name: 'p',
-        components: [
-          { name: 'lcaOwner', type: 'string' },
-          { name: 'tokenOwner', type: 'address' },
-          {
-            name: 'permit',
-            type: 'tuple',
-            internalType: PermitTransferFromInternalTypeName,
-            components: PermitTransferFromComponents,
-          },
-          { name: 'witness', type: 'bytes32' },
-          { name: 'witnessTypeString', type: 'string' },
-          { name: 'signature', type: 'bytes' },
-          { name: 'expectedWalletAddress', type: 'address' },
-        ],
-      },
-    ],
-    outputs: [],
-    stateMutability: 'nonpayable',
-  },
-] as const satisfies Abi;
-
 const depositFactoryCreateAndDeposit = depositFactoryABI.find(
   (
     item,
@@ -148,82 +73,5 @@ type DepositFactoryCreateAndDepositInput = NonNullable<
 >[0];
 export type CreateAndDepositPayload =
   AbiParameterToPrimitiveType<DepositFactoryCreateAndDepositInput>;
-
-/**
- * @see {@link https://github.com/agoric-labs/agoric-to-axelar-local/blob/main/packages/axelar-local-dev-cosmos/src/contracts/Wallet.sol Wallet.sol}
- * 1520650 2026-01-24
- */
-export const walletABI = [
-  {
-    type: 'constructor',
-    inputs: [
-      { name: 'gateway_', type: 'address' },
-      { name: 'gasReceiver_', type: 'address' },
-      { name: 'owner_', type: 'string' },
-    ],
-    stateMutability: 'payable',
-  },
-  {
-    type: 'event',
-    name: 'CallStatus',
-    anonymous: false,
-    inputs: [
-      { name: 'id', type: 'string', indexed: true },
-      { name: 'callIndex', type: 'uint256', indexed: true },
-      { name: 'target', type: 'address', indexed: true },
-      { name: 'methodSelector', type: 'bytes4', indexed: false },
-      { name: 'success', type: 'bool', indexed: false },
-    ],
-  },
-  {
-    type: 'event',
-    name: 'MulticallStatus',
-    anonymous: false,
-    inputs: [
-      { name: 'id', type: 'string', indexed: true },
-      { name: 'success', type: 'bool', indexed: false },
-      { name: 'totalCalls', type: 'uint256', indexed: false },
-    ],
-  },
-  {
-    type: 'event',
-    name: 'Received',
-    anonymous: false,
-    inputs: [
-      { name: 'sender', type: 'address', indexed: true },
-      { name: 'amount', type: 'uint256', indexed: false },
-    ],
-  },
-] as const satisfies Abi;
-
-/**
- * methods whose payload structure we need, even though
- * they are not public
- */
-const walletMulticallABI = [
-  {
-    type: 'function',
-    name: '_multicall',
-    inputs: [
-      {
-        type: 'tuple',
-        name: 'callMessage',
-        components: [
-          { name: 'id', type: 'string' },
-          {
-            name: 'calls',
-            type: 'tuple[]',
-            components: [
-              { name: 'target', type: 'address' },
-              { name: 'data', type: 'bytes' },
-            ],
-          },
-        ],
-      },
-    ],
-    outputs: [],
-    stateMutability: 'nonpayable',
-  },
-] as const satisfies Abi;
 
 export const walletCallMessageParams = walletMulticallABI[0].inputs ?? [];
